refactor(chat): extract pure entity/payload helpers to utils/chatHelpers.js
Chat.jsx carried ~250 lines of module-level pure helpers (conversation/
partner resolution, team-member checks, notification-payload parsing, team
detail merge). Move them verbatim into utils/chatHelpers.js and import the
public ones back, so the page component stays focused on orchestration.

Stage 1 of the Chat.jsx decomposition. No behaviour change: pure functions
moved as-is, internal-only helpers kept unexported. Chat.jsx 3543 -> 3296.
ESLint clean (pre-existing dedupeMessages dep warning unchanged), build green.
